Specify `testEnvironment: jest-environment-node` in jest.config.js
After bumping to the v5 package versions, tests would fail with the
following error:

```
● Test suite failed to run

  ReferenceError: TextEncoder is not defined

  > 1 | import * as pg from "pg";
      | ^
    2 |
    3 | export async function withPgPool<T = any>(
    4 |   cb: (pool: pg.Pool) => Promise<T>

    at Object.<anonymous> (node_modules/pg/lib/crypto/utils-webcrypto.js:22:21)
    at Object.<anonymous> (node_modules/pg/lib/crypto/utils.js:8:20)
    at Object.<anonymous> (node_modules/pg/lib/crypto/sasl.js:2:16)
    at Object.<anonymous> (node_modules/pg/lib/client.js:5:12)
    at Object.<anonymous> (node_modules/pg/lib/index.js:3:14)
    at Object.<anonymous> (__tests__/helpers.ts:1:1)
    at Object.<anonymous> (__tests__/schema.minimal_type.test.ts:2:1)
```

Based on [information from @SimenB](
  https://github.com/jsdom/jsdom/issues/2524#issuecomment-902027138
), it seems like you probably should not be using
`jest-environment-jsdom` to start with if you need access to
`TextEncoder` or `TextDecoder`.

Based on [this](
  https://stackoverflow.com/a/72369912/1137077
) answer on SO, I was able to verify that `@jest-environment node` at
the top of test files fixed the issue. However, it would seem more
logical to apply this as a global setting, given that postgraphile is
meant to run in a node context.

After setting `testEnvironment: jest-environment-node` in the global
config, I found that tests again started failing. However, the new
failures seem related to the v5 changes to plugins which is expected at
this stage.